requirements.txt
aiogram>=3.3,<4
aiosqlite>=0.19
python-dotenv>=1.0
